media-plugins/obs-backgroundremoval: use preset and add qt6 useflag
The default values in CMakeLists.txt fail to build, but they are
overwritten by different default values from CMakePresets.json, but
these presets need to be manually enabled.

But the preset also overwrite the binaryDir to a different directory, so
this needs to be set back to where gentoo expects it to be.

Also add a qt6 useflag to allow building with qt6.